requirements.txt
fdk>=0.1.75
pyjwt
cryptography
oci
requests